% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concatenate_python_codes}
\alias{concatenate_python_codes}
\title{Concatenate Code Files into one File}
\usage{
concatenate_python_codes(codes)
}
\arguments{
\item{codes}{a character vector of file names.}
}
\value{
The concatenated code.
}
\description{
To easily remove local imports and imports from __future__.
}
