% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{construct_path}
\alias{construct_path}
\title{Construct a (temporary) Path Where to Look for the Inst Files}
\usage{
construct_path(temporary = FALSE)
}
\arguments{
\item{temporary}{Use \code{\link{tempdir}()} as root for path construction?}
}
\value{
A character of length one naming the path constructed.
}
\description{
Construct a (temporary) Path Where to Look for the Inst Files
}
