% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excerptr-package.R
\docType{package}
\name{excerptr-package}
\alias{excerptr-package}
\title{Excerpt Structuring Comments from Your Code File and Set a Table of Contents}
\description{
This is an R interface to the 
Python package excerpts (\url{https://pypi.python.org/pypi/excerpts}).
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
\keyword{package}

