% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excerptr.R
\name{excerptr}
\alias{excerptr}
\title{Excerpt Structuring Comments and Set a Table of Contents.}
\usage{
excerptr(file_name, comment_character = "#", magic_character = "\%",
  output_path = "", allow_pep8 = TRUE, prefix = "", postfix = "",
  run_pandoc = TRUE, compile_latex = FALSE, pandoc_formats = "tex")
}
\arguments{
\item{file_name}{XXX}

\item{comment_character}{XXX}

\item{magic_character}{XXX}

\item{output_path}{XXX}

\item{allow_pep8}{XXX}

\item{prefix}{XXX}

\item{postfix}{XXX}

\item{run_pandoc}{XXX}

\item{compile_latex}{XXX}

\item{pandoc_formats}{XXX}
}
\value{
The status of the rPython::python.call() call to excerpts.
}
\description{
See somewhere on the web.
}
\examples{
root <- system.file(package = "excerptr")
test_files <- file.path(root, "excerpts", "tests", "files")
excerptr(file_name = file.path(test_files, "some_file.txt"),
         output_path = dirname(tempdir()), run_pandoc = TRUE,
         pandoc_formats = c("pdf"))
}

