% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_excerpts_path}
\alias{get_excerpts_path}
\title{Get the Excerpts Path Option}
\usage{
get_excerpts_path(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{set_excertps_path}()}.}
}
\value{
The status of the \pkg{rPython}\code{::\link[rPython]{python.call}}
call to excerpts.
}
\description{
Get the option "excerpts_path", and set it, if unset.
}
\examples{
options("excerpts_path" = NULL)
getOption("excertps_path")
excerptr:::get_excerpts_path() # internally calls set_excertps_path()
getOption("excertps_path")
}
\seealso{
\code{\link{set_excertps_path}}.
}
