% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_excertps_path}
\alias{set_excertps_path}
\title{Set the Excerpts Path Option}
\usage{
set_excertps_path(path = NULL, temporary = FALSE)
}
\arguments{
\item{path}{a fixed path. Equals \code{options(excerpts_path = path)}.}

\item{temporary}{Set a temporary path.}
}
\value{
a character of length one containing the (new) path set.
}
\description{
\pkg{excerptr} uses the option ("excerpts_path") to get the python code of
excerpts. This wrapper sets this option depending on whether you give a fixed
path, are developing the package or using the standard installation's code.
}
