% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excerptr-package.R
\docType{package}
\name{excerptr-package}
\alias{excerptr-package}
\title{Excerpt Structuring Comments from Your Code File and Set a Table of Contents}
\description{
This is just an R interface to the
python package excerpts (\url{https://pypi.python.org/pypi/excerpts}).
}
\details{
You will probably only want to use \code{\link{excerptr}}, see there for a
usage example. All other functions documented are rather meant for internal
use.
}
\keyword{package}
