% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_excerpts.R
\name{get_excerpts}
\alias{get_excerpts}
\title{Clone the Excerpts Code from Github}
\usage{
get_excerpts(directory = get_excerpts_path(), force = FALSE,
  remove_dot_git = TRUE)
}
\arguments{
\item{directory}{Where to clone to.}

\item{force}{Overwrite the directory?}

\item{remove_dot_git}{Remove the .git/ from the clone?}
}
\value{
The status of \code{git2r::clone}, or, if remove_dot_git is
TRUE, the \code{\link{capture.output}} thereof.
}
\description{
Clone the Excerpts Code from Github
}
\examples{
path <- file.path(tempdir(), "excerpts")
get_excerpts(directory = path, force = TRUE)
\dontrun{print(set_excertps_path(path = path))}
}
